added basic ability to attack

+++ SpaceBattle1/core/ship/armor/IArmor.cs
@@ -3,5 +3,5 @@
 public interface IArmor {
     String GetName();
     int GetBaseHitPoints();
-    double GetBaseDefense();
+    int GetBaseDefense();
 }
